requirements.txt
google-generativeai>=0.8.4
ollama>=0.4.7
pydantic>=2.10.6
asyncpg>=0.30.0
aiohttp>=3.11.12
python-dotenv>=1.0.1
sendgrid>=6.11.0
streamlit>=1.42.0
pandas>=2.2.3
